refactor(cli): split cli.ts into per-command modules
cli.ts was 1174 lines mixing the entry point, config model, setup
wizard, status/config/logs/doctor renderers, terminal chat bootstrap,
daemon outcome rendering, and the command router. Split into a cli/
folder (context, config, setup, status, config-view, logs, doctor, chat,
daemon, menu) with the router in index.ts.

src/cli.ts stays as the thin entry point — it keeps the shebang and the
must-run-first hidden subcommand dispatch (_mcp-launch / _lua-run), then
delegates to runCli(). bin/talon.js and package.json still point at
src/cli.ts, so the bin contract is unchanged. PKG_ROOT now resolves two
levels up from cli/context.ts (== repo root, same as before).

Verified: talon --version / --help / unknown-command suggester all work.
